feat : 설정 row button
